+++ Plone/src/docpool.event/docpool/event/content/dpevent.py
@@ -46,6 +46,8 @@ from z3c.relationfield.schema import RelationChoice
 from z3c.relationfield.schema import RelationList
 from zope import schema
 from zope.component import adapter
+from zope.component import getUtility
+from zope.interface import Interface
 from zope.interface import alsoProvides
 from zope.interface import implementer
 from zope.lifecycleevent.interfaces import IObjectAddedEvent
@@ -590,6 +592,10 @@ class DPEvent(Container, ContentBase):
             return
         return geometry.from_wkt(wkt)
 
+    def event_type_title(self):
+        vocab = getUtility(Interface, name='docpool.event.vocabularies.EventTypes')
+        return vocab(self).getTerm(self.EventType).title
+
 
 class ELANScenario(DPEvent):
     pass
@@ -613,6 +619,7 @@ def addLogEntry(obj):
         datetime.datetime.now(), long_format=1)
     entry[u'User'] = obj._getUserInfoString()
     entry[u'Status'] = obj.Status
+    entry[u'EventType'] = obj.EventType
     entry[u'Operation mode'] = obj.OperationMode
     entry[u'Alerting status'] = obj.AlertingStatus
     entry[u'Alerting note'] = obj.AlertingNote