+++ Plone/src/docpool.event/docpool/event/vocabularies.py
@@ -40,9 +40,10 @@ class EventTypesVocabulary(object):
 
     def __call__(self, context):
         values = [
-            (u'exercise', _(u'Exercise')),
-            (u'event', _(u'Event')),
-            (u'test', _(u'Test')),
+            (u'Emergency', _(u'Emergency')),
+            (u'Routine', _(u'Routine')),
+            (u'Exercise', _(u'Exercise')),
+            (u'Test', _(u'Test')),
             ]
         # value, token, title
         return SimpleVocabulary([SimpleTerm(i[0], i[0], i[1]) for i in values])